requirements.txt
# Core framework
dash>=2.17.0
dash-bootstrap-components>=1.6.0
flask>=3.0.0
gunicorn>=22.0.0

# Data & ML
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0
xgboost>=2.1.0
prophet>=1.1.5
statsmodels>=0.14.0
pmdarima>=2.0.4
shap>=0.45.0

# Visualization
plotly>=5.22.0

# API & networking
requests>=2.31.0

# Utilities
structlog>=24.1.0
python-dotenv>=1.0.0
holidays>=0.55

# Dev / testing
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-mock>=3.14.0
ruff>=0.5.0
mypy>=1.10.0
